Tacho and speedi proxies.

+++ soft/golang/slotman/services/impl/tacho/types.go
@@ -57,7 +57,7 @@ type Tacho struct {
 	// Response part.
 	//
 
-	Pin    byte      `json:",omitempty"`
+	Pin    int       `json:",omitempty"`
 	Active bool      `json:",omitempty"`
 	Time   time.Time `json:",omitempty"`
